requirements.txt
apturl==0.5.2
async-generator==1.10
attrs==19.3.0
bcrypt==3.1.7
black==22.6.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
h11==0.13.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
packaging==21.3
paramiko==2.6.0
pathspec==0.9.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2019.3
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
selenium==4.1.3
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.12
systemd-python==234
texttable==1.6.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==2.1.2
wsproto==1.1.0
xkit==0.0.0
zipp==1.0.0
